Move assembly tests to their own file and add some utilities (with tests) for assembly testing
