+++ tests/test_nodegraph.py
@@ -1060,539 +1060,3 @@ def test_traverse_linear_path_2():
     assert nodegraph.hash(contig[100:121]) in conns
     assert len(conns) == 1
 
-
-def test_traverse_linear_path_3_stopgraph():
-    contigfile = utils.get_test_data('simple-genome.fa')
-    contig = list(screed.open(contigfile))[0].sequence
-    print('contig len', len(contig))
-
-    K = 21
-
-    nodegraph = khmer.Nodegraph(K, 1e5, 4)
-    stopgraph = khmer.Nodegraph(K, 1e5, 4)
-
-    nodegraph.consume(contig)
-    nodegraph.count(contig[101:121] + 'G')  # will add another neighbor
-    print(nodegraph.neighbors(contig[101:122]))
-
-    degree_nodes = nodegraph.find_high_degree_nodes(contig)
-
-    assert len(degree_nodes) == 1
-    assert nodegraph.hash(contig[100:121]) in degree_nodes
-
-    stopgraph.count(contig[101:122])       # stop traversal - only adj to start
-
-    size, conns, visited = nodegraph.traverse_linear_path(contig[101:122],
-                                                          degree_nodes,
-                                                          stopgraph)
-
-    print(size, list(conns), len(visited))
-    assert size == 0
-    assert len(visited) == 0
-    assert len(conns) == 0
-
-
-def test_assemble_linear_path_1():
-    # assemble from beginning of contig, up until branch point
-    contigfile = utils.get_test_data('simple-genome.fa')
-    contig = list(screed.open(contigfile))[0].sequence
-    print('contig len', len(contig))
-
-    K = 21
-
-    nodegraph = khmer.Nodegraph(K, 1e5, 4)
-
-    nodegraph.consume(contig)
-    nodegraph.count(contig[101:121] + 'G')  # will add another neighbor
-
-    path = nodegraph.assemble_linear_path(contig[0:K])
-    len_path = len(path)
-
-    assert utils._equals_rc(path, contig[:len_path])
-
-
-def test_assemble_linear_path_1_rc():
-    # assemble from beginning of contig, up until branch point
-    # starting from rev comp
-    contigfile = utils.get_test_data('simple-genome.fa')
-    contig = list(screed.open(contigfile))[0].sequence
-    print('contig len', len(contig))
-
-    K = 21
-
-    nodegraph = khmer.Nodegraph(K, 1e5, 4)
-
-    nodegraph.consume(contig)
-    nodegraph.count(contig[101:121] + 'G')  # will add another neighbor
-
-    path = nodegraph.assemble_linear_path(revcomp(contig[0:K]))
-    len_path = len(path)
-
-    assert utils._equals_rc(path, contig[:len_path])
-
-
-def test_assemble_linear_path_2():
-    # assemble from branch point back to beginning of contig
-    contigfile = utils.get_test_data('simple-genome.fa')
-    contig = list(screed.open(contigfile))[0].sequence
-    print('contig len', len(contig))
-
-    K = 21
-
-    nodegraph = khmer.Nodegraph(K, 1e5, 4)
-
-    nodegraph.consume(contig)
-    nodegraph.count(contig[101:121] + 'G')  # will add another neighbor
-
-    path = nodegraph.assemble_linear_path(contig[100:100 + K])
-    len_path = len(path)
-
-    assert utils._equals_rc(path, contig[:len_path])
-
-
-def test_assemble_linear_path_2_rc():
-    # assemble from branch point back to beginning of contig
-    contigfile = utils.get_test_data('simple-genome.fa')
-    contig = list(screed.open(contigfile))[0].sequence
-    print('contig len', len(contig))
-
-    K = 21
-
-    nodegraph = khmer.Nodegraph(K, 1e5, 4)
-
-    nodegraph.consume(contig)
-    nodegraph.count(contig[101:121] + 'G')  # will add another neighbor
-
-    path = nodegraph.assemble_linear_path(revcomp(contig[100:100 + K]))
-    len_path = len(path)
-
-    assert utils._equals_rc(path, contig[:len_path])
-
-
-def test_assemble_linear_path_3():
-    # assemble entire contig, starting from wherever
-    contigfile = utils.get_test_data('simple-genome.fa')
-    contig = list(screed.open(contigfile))[0].sequence
-    print('contig len', len(contig))
-
-    K = 21
-
-    nodegraph = khmer.Nodegraph(K, 1e5, 4)
-
-    nodegraph.consume(contig)
-
-    for start in range(0, len(contig), 150):
-        path = nodegraph.assemble_linear_path(contig[start:start + K])
-        assert utils._equals_rc(path, contig), start
-
-
-def test_assemble_linear_path_4():
-    # assemble from branch point until end
-    contigfile = utils.get_test_data('simple-genome.fa')
-    contig = list(screed.open(contigfile))[0].sequence
-    print('contig len', len(contig))
-
-    K = 21
-
-    nodegraph = khmer.Nodegraph(K, 1e5, 4)
-
-    nodegraph.consume(contig)
-    nodegraph.count(contig[101:121] + 'G')  # will add another neighbor
-
-    path = nodegraph.assemble_linear_path(contig[101:101 + K])
-    len_path = len(path)
-
-    print('len path:', len_path)
-
-    # this is actually bad behavior; we're ignoring a branch.
-    # FIXME at some point.
-    assert utils._equals_rc(path, contig)
-    # should? be:
-    #    assert utils._equals_rc(path, contig[101:])
-
-
-def test_assemble_linear_path_5():
-    # assemble from end until branch point
-    contigfile = utils.get_test_data('simple-genome.fa')
-    contig = list(screed.open(contigfile))[0].sequence
-    print('contig len', len(contig))
-
-    K = 21
-
-    nodegraph = khmer.Nodegraph(K, 1e5, 4)
-
-    nodegraph.consume(contig)
-    nodegraph.count(contig[101:121] + 'G')  # will add another neighbor
-
-    path = nodegraph.assemble_linear_path(contig[-K:])
-    len_path = len(path)
-
-    print('len path:', len_path)
-
-    # this is actually bad behavior; we're ignoring a branch.
-    # FIXME at some point.
-    assert utils._equals_rc(path, contig)
-    # should? be:
-    #    assert utils._equals_rc(path, contig[101:])
-
-
-def test_assemble_linear_path_6():
-    # assemble from end until branch point
-    contigfile = utils.get_test_data('simple-genome.fa')
-    contig = list(screed.open(contigfile))[0].sequence
-    print('contig len', len(contig))
-
-    K = 21
-
-    nodegraph = khmer.Nodegraph(K, 1e5, 4)
-
-    nodegraph.consume(contig)
-    nodegraph.count('T' + contig[101:121])  # will add another neighbor
-
-    path = nodegraph.assemble_linear_path(contig[-K:])
-    len_path = len(path)
-
-    print('len path:', len_path)
-
-    assert utils._equals_rc(path, contig[101:])
-
-
-def test_assemble_linear_path_7():
-    # assemble from branch point until end
-    contigfile = utils.get_test_data('simple-genome.fa')
-    contig = list(screed.open(contigfile))[0].sequence
-    print('contig len', len(contig))
-
-    K = 21
-
-    nodegraph = khmer.Nodegraph(K, 1e5, 4)
-
-    nodegraph.consume(contig)
-    nodegraph.count('T' + contig[101:121])  # will add another neighbor
-
-    path = nodegraph.assemble_linear_path(contig[101:101 + K])
-    len_path = len(path)
-
-    print('len path:', len_path)
-
-    assert utils._equals_rc(path, contig[101:])
-
-
-def test_assemble_linear_path_8():
-    # assemble from branch point until end
-    contigfile = utils.get_test_data('simple-genome.fa')
-    contig = list(screed.open(contigfile))[0].sequence
-    print('contig len', len(contig))
-
-    K = 21
-
-    nodegraph = khmer.Nodegraph(K, 1e5, 4)
-    stop_bf = khmer.Nodegraph(K, 1e5, 4)
-
-    nodegraph.consume(contig)
-    nodegraph.count('T' + contig[101:121])  # will add another neighbor
-    stop_bf.count('T' + contig[101:121])    # ...that we will then ignore
-
-    path = nodegraph.assemble_linear_path(contig[101:101 + K], stop_bf)
-    len_path = len(path)
-
-    print('len path:', len_path)
-
-    assert utils._equals_rc(path, contig)
-
-
-def test_assemble_linear_path_9():
-    # assemble entire contig, ignoring branch point b/c of stop bf
-    contigfile = utils.get_test_data('simple-genome.fa')
-    contig = list(screed.open(contigfile))[0].sequence
-    print('contig len', len(contig))
-
-    K = 21
-
-    nodegraph = khmer.Nodegraph(K, 1e5, 4)
-    stop_bf = khmer.Nodegraph(K, 1e5, 4)
-
-    nodegraph.consume(contig)
-    nodegraph.count('T' + contig[101:121])  # will add another neighbor
-    stop_bf.count('T' + contig[101:121])    # ...that we will then ignore
-
-    path = nodegraph.assemble_linear_path(contig[-K:], stop_bf)
-    len_path = len(path)
-
-    print('len path:', len_path)
-
-    assert utils._equals_rc(path, contig)
-
-
-def test_assemble_linear_path_10():
-    # assemble up to branch point, and include introduced branch b/c
-    # of stop bf
-    contigfile = utils.get_test_data('simple-genome.fa')
-    contig = list(screed.open(contigfile))[0].sequence
-    print('contig len', len(contig))
-
-    K = 21
-
-    nodegraph = khmer.Nodegraph(K, 1e5, 4)
-    stop_bf = khmer.Nodegraph(K, 1e5, 4)
-
-    nodegraph.consume(contig)
-    nodegraph.count('T' + contig[101:121])  # will add another neighbor
-    stop_bf.count(contig[100:121])          # ...and block original path
-
-    path = nodegraph.assemble_linear_path(contig[-K:], stop_bf)
-    len_path = len(path)
-
-    print('len path:', len_path)
-
-    assert utils._equals_rc(path, 'T' + contig[101:])
-
-
-def test_assemble_linear_path_single_node():
-    # assemble single node.
-    contigfile = utils.get_test_data('simple-genome.fa')
-    contig = list(screed.open(contigfile))[0].sequence
-    print('contig len', len(contig))
-
-    K = 21
-
-    nodegraph = khmer.Nodegraph(K, 1e5, 4)
-
-    nodegraph.consume(contig)
-    nodegraph.count(contig[101:121] + 'G')  # will add another neighbor after
-    nodegraph.count('T' + contig[100:120])  # ...and before.
-
-    path = nodegraph.assemble_linear_path(contig[100:121])
-    len_path = len(path)
-
-    print('len path:', len_path)
-
-    assert utils._equals_rc(path, contig[100:121])
-
-
-def test_assemble_linear_path_single_node_interrupted():
-    # assemble single node.
-    contigfile = utils.get_test_data('simple-genome.fa')
-    contig = list(screed.open(contigfile))[0].sequence
-    print('contig len', len(contig))
-
-    K = 21
-
-    nodegraph = khmer.Nodegraph(K, 1e5, 4)
-
-    nodegraph.consume(contig)
-    nodegraph.consume(contig[:110] + 'G')  # will add another neighbor/middle
-
-    path = nodegraph.assemble_linear_path(contig[100:121])
-    len_path = len(path)
-
-    print('len path:', len_path)
-
-    assert utils._equals_rc(path, contig)       # this is bad behavior...
-
-
-def test_assemble_labeled_paths():
-    # assemble entire contig, ignoring branch point b/c of labels
-    contigfile = utils.get_test_data('simple-genome.fa')
-    contig = list(screed.open(contigfile))[0].sequence
-    print('contig len', len(contig))
-
-    K = 21
-
-    nodegraph = khmer.Nodegraph(K, 1e5, 4)
-    lh = khmer._GraphLabels(nodegraph)
-
-    nodegraph.consume(contig)
-    nodegraph.count(contig[100:120] + 'T')  # will add another neighbor
-
-    print(contig[100:125])
-
-    hdn = nodegraph.find_high_degree_nodes(contig)
-    lh.label_across_high_degree_nodes(contig, hdn, 1)
-
-    path = lh.assemble_labeled_path(contig[:K])
-    path = path[0]  # @CTB
-    len_path = len(path)
-
-    print('len path:', len_path)
-
-    assert utils._equals_rc(path, contig)
-
-
-def test_assemble_labeled_paths_2():
-    # assemble entire contig + branch point b/c of labels
-    contigfile = utils.get_test_data('simple-genome.fa')
-    contig = list(screed.open(contigfile))[0].sequence
-    print('contig len', len(contig))
-
-    K = 21
-
-    nodegraph = khmer.Nodegraph(K, 1e5, 4)
-    lh = khmer._GraphLabels(nodegraph)
-
-    nodegraph.consume(contig)
-    branch = contig[:120] + 'TGATGGACAG'
-    nodegraph.consume(branch)  # will add a branch
-
-    hdn = nodegraph.find_high_degree_nodes(contig)
-    hdn += nodegraph.find_high_degree_nodes(branch)
-    print(list(hdn))
-    lh.label_across_high_degree_nodes(contig, hdn, 1)
-    lh.label_across_high_degree_nodes(branch, hdn, 2)
-    print(lh.get_tag_labels(list(hdn)[0]))
-
-    paths = lh.assemble_labeled_path(contig[:K])
-    print([len(x) for x in paths])
-    len_path = len(paths)
-
-    print('len path:', len_path)
-
-    found = False
-    for path in paths:
-        if utils._equals_rc(path, contig):
-            found = True
-            break
-    assert found
-
-    found = False
-    for path in paths:
-        if utils._equals_rc(path, branch):
-            found = True
-            break
-    assert found
-
-
-def test_assemble_labeled_paths_3():
-    # assemble entire contig + branch points b/c of labels
-    contigfile = utils.get_test_data('simple-genome.fa')
-    contig = list(screed.open(contigfile))[0].sequence
-    print('contig len', len(contig))
-
-    K = 21
-
-    nodegraph = khmer.Nodegraph(K, 1e5, 4)
-    lh = khmer._GraphLabels(nodegraph)
-
-    nodegraph.consume(contig)
-    branch = contig[:120] + 'TGATGGACAG'
-    nodegraph.consume(branch)  # will add a branch
-    branch2 = contig[:120] + 'GCGGATGGATGGAGCCGAT'
-    nodegraph.consume(branch2)  # will add a third branch
-
-    hdn = nodegraph.find_high_degree_nodes(contig)
-    hdn += nodegraph.find_high_degree_nodes(branch)
-    hdn += nodegraph.find_high_degree_nodes(branch2)
-    print(list(hdn))
-    lh.label_across_high_degree_nodes(contig, hdn, 1)
-    lh.label_across_high_degree_nodes(branch, hdn, 2)
-    lh.label_across_high_degree_nodes(branch2, hdn, 3)
-    print(lh.get_tag_labels(list(hdn)[0]))
-
-    paths = lh.assemble_labeled_path(contig[:K])
-    print([len(x) for x in paths])
-    len_path = len(paths)
-
-    print('len path:', len_path)
-
-    found = False
-    for path in paths:
-        if utils._equals_rc(path, contig):
-            found = True
-            break
-    assert found
-
-    found = False
-    for path in paths:
-        if utils._equals_rc(path, branch):
-            found = True
-            break
-    assert found
-
-    found = False
-    for path in paths:
-        if utils._equals_rc(path, branch2):
-            found = True
-            break
-    assert found
-
-
-def test_assemble_labeled_paths_4():
-    # assemble entire contig + branch points b/c of labels; start from end
-    contigfile = utils.get_test_data('simple-genome.fa')
-    contig = list(screed.open(contigfile))[0].sequence
-    print('contig len', len(contig))
-
-    K = 21
-
-    nodegraph = khmer.Nodegraph(K, 1e5, 4)
-    lh = khmer._GraphLabels(nodegraph)
-
-    nodegraph.consume(contig)
-    branch = 'TGATGGACAG' + contig[120:]
-    nodegraph.consume(branch)  # will add a branch
-    branch2 = 'GCGGATGGATGGAGCCGAT' + contig[120:]
-    nodegraph.consume(branch2)  # will add a third branch
-
-    hdn = nodegraph.find_high_degree_nodes(contig)
-    hdn += nodegraph.find_high_degree_nodes(branch)
-    hdn += nodegraph.find_high_degree_nodes(branch2)
-    print(list(hdn))
-    lh.label_across_high_degree_nodes(contig, hdn, 1)
-    lh.label_across_high_degree_nodes(branch, hdn, 2)
-    lh.label_across_high_degree_nodes(branch2, hdn, 3)
-    print(lh.get_tag_labels(list(hdn)[0]))
-
-    paths = lh.assemble_labeled_path(contig[-K:])
-    print([len(x) for x in paths])
-    len_path = len(paths)
-
-    print('len path:', len_path)
-
-    found = False
-    for path in paths:
-        if utils._equals_rc(path, contig):
-            found = True
-            break
-    assert found
-
-    found = False
-    for path in paths:
-        if utils._equals_rc(path, branch):
-            found = True
-            break
-    assert found
-
-    found = False
-    for path in paths:
-        if utils._equals_rc(path, branch2):
-            found = True
-            break
-    assert found
-
-
-def test_assemble_labeled_paths_5():
-    # assemble entire contig + one of two paths through a bubble
-    contigfile = utils.get_test_data('simple-genome.fa')
-    contig = list(screed.open(contigfile))[0].sequence
-    print('contig len', len(contig))
-
-    K = 21
-
-    nodegraph = khmer.Nodegraph(K, 1e5, 4)
-    lh = khmer._GraphLabels(nodegraph)
-
-    nodegraph.consume(contig)
-    contig2 = contig[:200] + 'G' + contig[201:]
-    nodegraph.consume(contig2)
-
-    hdn = nodegraph.find_high_degree_nodes(contig)
-    assert len(hdn) == 2
-    lh.label_across_high_degree_nodes(contig, hdn, 1)
-
-    path = lh.assemble_labeled_path(contig[:K])
-    path = path[0]  # @CTB
-    len_path = len(path)
-
-    print('len path:', len_path)
-
-    assert utils._equals_rc(path, contig)
